chore: organize routes
