+++ api/test/routes/index_test.dart
@@ -4,7 +4,7 @@ import 'package:dart_frog/dart_frog.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
-import '../../../routes/public/index.dart' as route;
+import '../../routes/index.dart' as route;
 
 class _MockRequestContext extends Mock implements RequestContext {}
